Write basic tests for customer model
First of all, get rid of that shitty Java library called unittest2 and
get some pytest going.  The tests are to be run in the root dir of the
project by enabling the virtual environment if necessary and then
invoking:

  $ pytest

This reads the provided pytest.ini file to configure the environment.  I
personally like some verbosity, hence:

  $ pytest -v

Currently the following tests have been implemented on the Customer
model:

 - Check customer details after object creation

 - Check whether submitting NULL values to NOT NULL fields raise
   ValueError.

 - Check whether submitting empty strings to NOT NULL fields raise
   ValueError.

 - Check whether email invalidation raises a ValidationError.

 - Check exepcted full name is returned in the __str__() method of
   Customer.

Enforce the necessary checks and exceptions on the submitted data to
satisfy the tests.